Move game logic into a shared folder
