+++ shared/interfaces.ts
